requirements.txt
pygame==2.4.0
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.6
numpy==1.24.3
Pillow==10.0.0
